+++ java/src/c22485282/GradientShape.java
@@ -1,4 +1,4 @@
-package example;
+package c22485282;
 
 import processing.core.PApplet;
