+++ java/src/c22485282/Worm.java
@@ -1,4 +1,4 @@
-package example;
+package c22485282;
 
 public class Worm {
